requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
Markdown==3.9
packaging==25.0
psycopg==3.2.11
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
